+++ example/http/server.js
@@ -25,6 +25,7 @@ jwt.setKey('12345');
 console.log('A TOKEN => ', jwt.generateAccessToken());
 app.use(jwt.isAuthorized(jwt));
 app.use(jwt.autoEncryptRequest(jwt));
+app.use(jwt.autoDecryptRequest(jwt));
 
 // Configure app to angular and all bower_components
 app.use('/public', express.static(base + '/public'));
@@ -33,7 +34,7 @@ app.use('/bower_components',  express.static(base + '/bower_components'));
 // Request to connect
 // email = 'toto' and pwd == 'aaaa'
 app.post('/login', function(req, res) {
-
+  console.log('B in route =>', req.body);
   if (req.body.email == 'toto' && req.body.pwd == 'aaaa') {
     console.log("connect success");
 
@@ -44,7 +45,6 @@ app.post('/login', function(req, res) {
     res.status(400).json({ message : 'connect Failed' });
   }
 });
-
 
 app.get('/home', function(req, res, next) {
   res.status(200).jsonp({ message :' welcome to the home' });